# frontend/openapi/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "description" : "Modular-monolith manufacturing ERP — double-entry ledger, inventory, procure-to-pay, order-to-cash, manufacturing and reporting.",
    "title" : "Manufacturing ERP API",
    "version" : "0.0.1"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
  "paths" : {
...
    "/api/masterdata/items" : {
      "get" : {
        "operationId" : "listItems",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ItemResponse"
                  }
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "master-data-controller" ]
      },
      "post" : {
        "operationId" : "createItem",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateItemRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ItemResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "master-data-controller" ]
      }
    },
    "/api/masterdata/items/by-sku" : {
      "get" : {
        "operationId" : "getItemBySku",
        "parameters" : [ {
          "in" : "query",
          "name" : "sku",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ItemResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "master-data-controller" ]
      }
    },
    "/api/masterdata/items/{id}" : {
      "get" : {
        "operationId" : "getItem",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ItemResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "master-data-controller" ]
      }
    },
    "/api/masterdata/locations" : {
      "get" : {
        "operationId" : "listLocations",
        "parameters" : [ {
          "in" : "query",
          "name" : "warehouseId",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/LocationResponse"
                  }
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "master-data-controller" ]
      },
      "post" : {
        "operationId" : "createLocation",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateLocationRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/LocationResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "master-data-controller" ]
      }
    },
    "/api/masterdata/locations/{id}" : {
      "get" : {
        "operationId" : "getLocation",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/LocationResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "master-data-controller" ]
      }
    },
    "/api/masterdata/partners" : {
      "get" : {
        "operationId" : "listPartners",
        "parameters" : [ {
          "in" : "query",
          "name" : "vendor",
          "required" : false,
          "schema" : {
            "type" : "boolean"
          }
        }, {
          "in" : "query",
          "name" : "customer",
          "required" : false,
          "schema" : {
            "type" : "boolean"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/PartnerResponse"
                  }
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "master-data-controller" ]
      },
      "post" : {
        "operationId" : "createPartner",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreatePartnerRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PartnerResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "master-data-controller" ]
      }
    },
    "/api/masterdata/partners/{id}" : {
      "get" : {
        "operationId" : "getPartner",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PartnerResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "master-data-controller" ]
      }
    },
    "/api/masterdata/warehouses" : {
      "get" : {
        "operationId" : "listWarehouses",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/WarehouseResponse"
                  }
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "master-data-controller" ]
      },
      "post" : {
        "operationId" : "createWarehouse",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateWarehouseRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WarehouseResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "master-data-controller" ]
      }
    },
    "/api/payments" : {
      "get" : {
        "operationId" : "listPayments",
        "parameters" : [ {
          "in" : "query",
          "name" : "direction",
          "required" : false,
          "schema" : {
            "type" : "string",
            "enum" : [ "IN", "OUT" ]
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/PaymentResponse"
                  }
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "payments-controller" ]
      }
    },
    "/api/payments/in" : {
      "post" : {
        "operationId" : "payIn",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PayInRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PaymentResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "payments-controller" ]
      }
    },
    "/api/payments/out" : {
      "post" : {
        "operationId" : "payOut",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PayOutRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PaymentResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "payments-controller" ]
      }
    },
    "/api/payments/{id}" : {
      "get" : {
        "operationId" : "getPayment",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PaymentResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "payments-controller" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "AllocationLine" : {
        "type" : "object",
        "properties" : {
          "amount" : {
            "type" : "string"
          },
          "billId" : {
            "type" : "integer",
            "format" : "int64"
          }
        },
        "required" : [ "amount", "billId" ]
      },
      "AllocationView" : {
        "type" : "object",
        "properties" : {
          "amount" : {
            "type" : "string"
          },
          "documentId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "id" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
...
      "CreateItemRequest" : {
        "type" : "object",
        "properties" : {
          "itemType" : {
            "type" : "string",
            "enum" : [ "RAW", "WIP", "FINISHED", "SERVICE" ]
          },
          "name" : {
            "type" : "string",
            "minLength" : 1
          },
          "reorderPoint" : {
            "type" : "string"
          },
          "reorderQty" : {
            "type" : "string"
          },
          "sku" : {
            "type" : "string",
            "minLength" : 1
          },
          "standardCost" : {
            "type" : "string"
          },
          "stocked" : {
            "type" : "boolean"
          },
          "uom" : {
            "type" : "string",
            "minLength" : 1
          }
        },
        "required" : [ "itemType", "name", "sku", "uom" ]
      },
...
      "CreateLocationRequest" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "string",
            "minLength" : 1
          },
          "locationType" : {
            "type" : "string",
            "enum" : [ "STOCK", "RECEIVING", "SHIPPING", "PRODUCTION_WIP", "SCRAP", "VENDOR", "CUSTOMER", "INVENTORY_LOSS" ]
          },
          "warehouseId" : {
            "type" : "integer",
            "format" : "int64"
          }
        },
        "required" : [ "code", "locationType", "warehouseId" ]
      },
      "CreatePartnerRequest" : {
        "type" : "object",
        "properties" : {
          "apAccountCode" : {
            "type" : "string"
          },
          "arAccountCode" : {
            "type" : "string"
          },
          "code" : {
            "type" : "string",
            "minLength" : 1
          },
          "customer" : {
            "type" : "boolean"
          },
          "name" : {
            "type" : "string",
            "minLength" : 1
          },
          "paymentTermsDays" : {
            "type" : "integer",
            "format" : "int32"
          },
          "taxId" : {
            "type" : "string"
          },
          "vendor" : {
            "type" : "boolean"
          }
        },
        "required" : [ "code", "name" ]
      },
...
      "CreateWarehouseRequest" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "string",
            "minLength" : 1
          },
          "name" : {
            "type" : "string",
            "minLength" : 1
          }
        },
        "required" : [ "code", "name" ]
      },
...
      "ItemResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "itemType" : {
            "type" : "string",
            "enum" : [ "RAW", "WIP", "FINISHED", "SERVICE" ]
          },
          "name" : {
            "type" : "string"
          },
          "reorderPoint" : {
            "type" : "string"
          },
          "reorderQty" : {
            "type" : "string"
          },
          "sku" : {
            "type" : "string"
          },
          "standardCost" : {
            "type" : "string"
          },
          "stocked" : {
            "type" : "boolean"
          },
          "uom" : {
            "type" : "string"
          }
        }
      },
...
      "LocationResponse" : {
        "type" : "object",
        "properties" : {
          "active" : {
            "type" : "boolean"
          },
          "code" : {
            "type" : "string"
          },
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "locationType" : {
            "type" : "string",
            "enum" : [ "STOCK", "RECEIVING", "SHIPPING", "PRODUCTION_WIP", "SCRAP", "VENDOR", "CUSTOMER", "INVENTORY_LOSS" ]
          },
          "warehouseId" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
...
      "PartnerResponse" : {
        "type" : "object",
        "properties" : {
          "active" : {
            "type" : "boolean"
          },
          "apAccountCode" : {
            "type" : "string"
          },
          "arAccountCode" : {
            "type" : "string"
          },
          "code" : {
            "type" : "string"
          },
          "customer" : {
            "type" : "boolean"
          },
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "name" : {
            "type" : "string"
          },
          "paymentTermsDays" : {
            "type" : "integer",
            "format" : "int32"
          },
          "taxId" : {
            "type" : "string"
          },
          "vendor" : {
            "type" : "boolean"
          }
        }
      },
      "PayInRequest" : {
        "type" : "object",
        "properties" : {
          "allocations" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ReceiptAllocationLine"
            }
          },
          "amount" : {
            "type" : "string"
          },
          "partnerId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "postingDate" : {
            "type" : "string",
            "format" : "date"
          }
        },
        "required" : [ "allocations", "amount", "partnerId" ]
      },
      "PayOutRequest" : {
        "type" : "object",
        "properties" : {
          "allocations" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/AllocationLine"
            }
          },
          "amount" : {
            "type" : "string"
          },
          "partnerId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "postingDate" : {
            "type" : "string",
            "format" : "date"
          }
        },
        "required" : [ "allocations", "amount", "partnerId" ]
      },
      "PaymentResponse" : {
        "type" : "object",
        "properties" : {
          "allocations" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/AllocationView"
            }
          },
          "amount" : {
            "type" : "string"
          },
          "direction" : {
            "type" : "string"
          },
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "journalEntryId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "partnerId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "payNumber" : {
            "type" : "string"
          },
          "postedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "postingDate" : {
            "type" : "string",
            "format" : "date"
          },
          "status" : {
            "type" : "string"
          }
        }
      },
...
      "ReceiptAllocationLine" : {
        "type" : "object",
        "properties" : {
          "amount" : {
            "type" : "string"
          },
          "invoiceId" : {
            "type" : "integer",
            "format" : "int64"
          }
        },
        "required" : [ "amount", "invoiceId" ]
      },
...
      "WarehouseResponse" : {
        "type" : "object",
        "properties" : {
          "active" : {
            "type" : "boolean"
          },
          "code" : {
            "type" : "string"
          },
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "name" : {
            "type" : "string"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  }
}